Release Energy Device Types
