+++ .vscode/tasks.json
@@ -137,15 +137,15 @@
         "--verbose",
         "--coverage",
         "--maxWorkers=100%",
-        "src/base.test.ts",
-        "src/laundryWasher.test.ts",
-        "src/laundryDryer.test.ts",
-        "src/roboticVacuumCleaner.test.ts",
-        "src/evse.test.ts",
-        "src/waterHeater.test.ts",
-        "src/solarPower.test.ts",
-        "src/batteryStorage.test.ts",
-        "src/heatPump.test.ts"
+        "src/devices/base.test.ts",
+        "src/devices/laundryWasher.test.ts",
+        "src/devices/laundryDryer.test.ts",
+        "src/devices/roboticVacuumCleaner.test.ts",
+        "src/devices/evse.test.ts",
+        "src/devices/waterHeater.test.ts",
+        "src/devices/solarPower.test.ts",
+        "src/devices/batteryStorage.test.ts",
+        "src/devices/heatPump.test.ts"
       ],
       "group": {
         "kind": "test",
