+++ src/devices/export.ts
@@ -1,8 +1,9 @@
 // Single class device types
-export * from '../roboticVacuumCleaner.js';
-export * from '../laundryWasher.js';
-export * from '../laundryDryer.js';
-export * from '../waterHeater.js';
-export * from '../evse.js';
-export * from '../solarPower.js';
-export * from '../batteryStorage.js';
+export * from './roboticVacuumCleaner.js';
+export * from './laundryWasher.js';
+export * from './laundryDryer.js';
+export * from './waterHeater.js';
+export * from './evse.js';
+export * from './solarPower.js';
+export * from './batteryStorage.js';
+export * from './heatPump.js';
